Move remaining test cases
